Moved and changed most of the .h files
