+++ SluaghEngine/Core/Core.vcxproj
@@ -141,11 +141,13 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\includes\Core\TransformManager.cpp" />
+    <ClCompile Include="AudioManager.cpp" />
     <ClCompile Include="Engine.cpp" />
     <ClCompile Include="EntityManager.cpp" />
     <ClCompile Include="RenderableManager.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\includes\Core\AudioManager.h" />
     <ClInclude Include="..\includes\Core\Engine.h" />
     <ClInclude Include="..\includes\Core\EntityManager.h" />
     <ClInclude Include="..\includes\Core\TransformManager.h" />
