+++ SluaghEngine/Test/Test.vcxproj.filters
@@ -45,6 +45,9 @@
     <ClCompile Include="RenderableManagerTest.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="AudioTest.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="EntityManagerTest.h">
@@ -72,12 +75,15 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="TransformManagerTest.h">
-	 <Filter>Header Files</Filter>
-	 </ClInclude>
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="ObjLoaderTest.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="RenderableManagerTest.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="AudioTest.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
