+++ SluaghEngine/Test/rawLoaderEntries.txt
@@ -1,3 +1,5 @@
 test.txt
 Asset/test.objtest
 Asset/Placeholder_level.obj
+Asset/Canary.wav
+Asset/Cout.wav
